# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/responses/compact": {
      "post": {
        "summary": "压缩对话 (OpenAI Responses API)",
        "deprecated": false,
        "description": "OpenAI Responses API，用于对长对话进行 compaction。",
        "operationId": "compactResponse",
        "tags": [
          "OpenAI格式(Responses)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesCompactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功压缩对话",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesCompactionResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "上游模型最终不可用。网关隐藏上游账户和原始错误信息，并按客户端协议返回统一的当前模型不可用错误。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUnavailableResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModelUnavailableResponse": {
        "description": "请求已到达上游但最终失败时的公开错误。网关不会暴露上游账户、鉴权、余额或原始响应；不同协议保留各自的外层错误结构。",
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "当前模型不可用 (request id: req_example)"
                  },
                  "type": {
                    "type": "string",
                    "example": "server_error"
                  },
                  "param": {
                    "type": "string",
                    "example": ""
                  },
                  "code": {
                    "type": "string",
                    "example": "model_unavailable"
                  }
                },
                "required": [
                  "message",
                  "type",
                  "code"
                ]
              }
            },
            "required": [
              "error"
            ]
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "model_unavailable"
              },
              "message": {
                "type": "string",
                "example": "当前模型不可用"
              },
              "data": {
                "nullable": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "error"
              },
              "error": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "model_unavailable"
                  },
                  "message": {
                    "type": "string",
                    "example": "当前模型不可用 (request id: req_example)"
                  }
                },
                "required": [
                  "type",
                  "message"
                ]
              }
            },
            "required": [
              "type",
              "error"
            ]
          }
        ]
      },
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "description": "提示词 Token 数"
          },
          "completion_tokens": {
            "type": "integer",
            "description": "补全 Token 数"
          },
          "total_tokens": {
            "type": "integer",
            "description": "总 Token 数"
          },
          "prompt_tokens_details": {
            "type": "object",
            "properties": {
              "cached_tokens": {
                "type": "integer"
              },
              "text_tokens": {
                "type": "integer"
              },
              "audio_tokens": {
                "type": "integer"
              },
              "image_tokens": {
                "type": "integer"
              }
            }
          },
          "completion_tokens_details": {
            "type": "object",
            "properties": {
              "text_tokens": {
                "type": "integer"
              },
              "audio_tokens": {
                "type": "integer"
              },
              "reasoning_tokens": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "ResponsesCompactionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string",
            "example": "response.compaction"
          },
          "created_at": {
            "type": "integer"
          },
          "output": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {}
            }
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "error": {
            "type": "object",
            "properties": {}
          }
        }
      },
      "ResponsesCompactionRequest": {
        "type": "object",
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string"
          },
          "input": {
            "description": "输入内容，可以是字符串或消息数组",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {}
                }
              }
            ]
          },
          "instructions": {
            "type": "string"
          },
          "previous_response_id": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}